+++ src/discaccess_emulator.h
@@ -1,1 +1,1 @@
-void discaccess_emulator_init( uint8_t instance , int address);
+void discaccess_emulator_init( uint8_t instance , uint8_t address);
