+++ src/ram_emulator.h
@@ -1,5 +1,5 @@
 
-void rampage_emulator_init( uint8_t instance , int address);
-void rambyte_emulator_init( uint8_t instance , int address);
+void rampage_emulator_init( uint8_t instance , uint8_t address);
+void rambyte_emulator_init( uint8_t instance , uint8_t address);
 
 void ram_emulator_page_restore(void);
